Better curses prog, but list-graph problem
Only 1 class for curses progs, linked as Unix pipes. But, the problem is
that the matching currently returns a list of nodes with highlights, not
the same type of graph as the input. It needs some kind of system of
selecting a subgraph and then printing only the selected nodes.

	modified:   curses_menu.py