Processing the lists of options instead of node graph
The node graph is converted into the list of options. The curses
programs work as a Unix filter over these lists.

Not sure of it works for interesting applications: more graph selections
etc. But, if there is more graph works, it should be done with some
proper graph library and language, some graph SQL.

Here, it is a nice CSS-like language of selectors. And it's only one
class. It's a single-class library. The graph-sql should be something
like that but more.

Should work for: OPC client, OpenAPI jsons, CMake build jsons, simple
web scrapping.

	modified:   curses_menu.py