OK, it's clunky but writing OPC works
I need to really read the asyncua lib and the OPC standard. There are
smart things. But the asyncua code (uabrowse, uaread etc tools) is
really clunky. It does not use the handy methods like node.write_value
etc. The translation and guessing of types is also a one-off thing.
The coroutines are sometimes confusing. Seems like not all of them are
needed. But the point is that the standard is smart and the library
delivers it.

	modified:   curses_menu.py
	modified:   get_opcua_datapoints.py